Get session test created
